requirements.txt
# Fetching
requests>=2.31.0
beautifulsoup4>=4.12.0   # text extraction only (no selector parsing)

# LLM extraction
anthropic>=0.40.0

# Telegram
telethon>=1.34.0

# Utils
python-dotenv>=1.0.0
